chore: move offline RPC server to a separate file
