+++ src/tool/offline_server/mod.rs
@@ -1,5 +1,5 @@
 // Copyright 2019-2024 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
-pub mod main;
-mod offline_server;
-pub mod subcommands;
+
+mod server;
+pub use server::start_offline_server;
